+++ CHANGELOG.markdown
@@ -1,6 +1,9 @@
 next
 ----
 * Add a `Generic` default implementation for `Wrapped`
+* Revamp `Setup.hs` to use `cabal-doctest`. This makes it build
+  with `Cabal-1.25`, and makes the `doctest`s work with `cabal new-build` and
+  sandboxes.
 
 4.15.1
 ----