+++ CHANGELOG.markdown
@@ -11,6 +11,7 @@ next
 * Revamp `Setup.hs` to use `cabal-doctest`. This makes it build
   with `Cabal-1.25`, and makes the `doctest`s work with `cabal new-build` and
   sandboxes.
+* Added `throwing_` to `Control.Monad.Error.Lens` and `Control.Exception.Lens`
 
 4.15.1
 ----